Conflicts:
	app/models/user.rb

+++ app/workers/facebook_friends_worker.rb
@@ -7,13 +7,20 @@ class FacebookFriendsWorker
     user = User.where(id: user_id).first
     return if user.nil?
     auth = Authentication.where(provider: 'facebook', user_id: user.id).last
+    return if auth.nil?
     graph = Koala::Facebook::API.new(auth.token)
-    profile = graph.get_object("me")
+    begin
+      profile = graph.get_object("me")
+    rescue Koala::Facebook::AuthenticationError => e
+      auth.destroy
+      return
+    end
     me = FacebookIdentity.new().update_from_facebook(profile)
     me.user_id = user.id
     me = me.merge
     return if me.refreshed_at > 5.minutes.ago
     ActiveRecord::Base.transaction do
+      count = 0
       me.update!(:refreshed_at => Time.now)
       me.friendships.where(:friend_type => 'FacebookIdentity').destroy_all
       result = graph.get_connections("me", "friends", :fields=>"name,id,picture.width(256).height(256)") || []
@@ -23,9 +30,14 @@ class FacebookFriendsWorker
           fid = fid.merge
           fs = Friendship.new( identity: me, friend: fid )
           fs = fs.merge
+          count = count + 1
         end
         result = result.next_page || []
       end while not result.empty?
+      logger.info "Refreshed user #{user_id}'s #{count} facebook friends"
     end
+    # Touch all changed friendships outside of transaction in case a sync happened while isolated
+    me.friendships.where(:friend_type => 'FacebookIdentity')
+                  .where('updated_at >= ?', me.refreshed_at).update_all(updated_at: Time.now)
   end
 end
